requirements.txt
asyncio==3.4.3
pyaes==1.6.1
pyasn1==0.4.7
PySocks==1.7.0
rsa==4.0
Telethon==1.10.3
